Implement downloader in python too.
Passing cookies to wget on the command line is a security risk. On a
shared machine, other users can see your full command line. Passing it
in using the --load-cookies option is too tedious—the file format
required is archaic and hard to replicate by hand. So, we simply
implement the downloader in python too. In any case, this makes for a
more cohesive user experience.